# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/posts/{post}": {
            "delete": {
                "tags": [
                    "Посты"
                ],
                "summary": "Удаление поста",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "Пост был удалён"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Посты"
                ],
                "summary": "Редактирование поста",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Sometitle"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Lorem ipsum dolor sit amet consectetur adipisicing elit. Ex minus, impedit commodi nobis tempore dolores. Cupiditate deserunt temporibus, sunt quia explicabo veritatis assumenda voluptatibus quas, autem ipsa tempora magni aliquam dolorum veniam sapiente expedita recusandae aperiam soluta enim saepe reiciendis odio consequatur dolores omnis. Aliquid excepturi repellat, hic recusandae minima autem enim consectetur fuga maiores quia quidem tempore atque est ea cumque voluptatibus odit veniam ducimus. Expedita omnis molestiae, eos aliquid error saepe nesciunt magnam magni nobis earum, sit nostrum sed possimus dolor eveniet iure eligendi quam corporis cupiditate laboriosam. Molestiae suscipit unde aperiam ducimus ipsa eius voluptate quia? Nesciunt asperiores temporibus voluptates, id voluptatum, ullam accusantium ipsum eveniet totam ut impedit non aut, dolorum corporis hic cum similique exercitationem dolorem consequatur magnam cupiditate officiis. Quidem at architecto sit illum deserunt quas earum cum, fugit laboriosam excepturi similique, quod quasi distinctio, facere minus tempora dolorem! Officia illum quae inventore quidem hic omnis minima voluptate, fugit deleniti assumenda fuga repudiandae possimus, amet nulla, facilis dignissimos. Eveniet culpa neque dolores nulla sunt suscipit, voluptatum laboriosam nam libero, magni omnis ut fuga, consectetur soluta ipsum sequi labore. Dignissimos adipisci autem laborum quo exercitationem, asperiores eum rem sunt eveniet architecto, quae odit? Necessitatibus, a."
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "SomeTitle"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "SomeContent"
                                                },
                                                "category": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "CategoryTitle"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/posts/{post}/likes": {
            "post": {
                "tags": [
                    "Лайки"
                ],
                "summary": "Добавление лайка к посту",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "post_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}